# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto API
  description: API for WasaPhoto
  version: 1.0.0
...
components:
  schemas:
...
    FollowRequest:
      description: 'It which user we want to follow'
      type: object
      properties:
        userID:
          description: "UserID of the person"
          type: integer
          format: int64
      example:
        userID: 123
      required:
        - userID
    BanRequest:
      description: 'It which user we want to ban'
      type: object
      properties:
        userID:
          description: "UserID of the person who we want to ban"
          type: integer
          format: int64
      example:
        userID: 123
      required:
        - userID
    Comment:
      description: 'It contains the text data for the comment on an image'
      type: object
      properties:
        text:
          description: "Text we want to be the comment"
          type: string
          pattern: "^.*?$"
          minLength: 2 
          maxLength: 20
      example:
        text: 'Nice picture'
      required: 
        - text
...
  responses:
...
    NotFoundResponseUser:
      description: "User not found"
    NoContentResponse:
      description: "No content found"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
...
  /users/{userID}/follow: 
    post:
      security:
        - bearerAuth: []
      summary: Follow a user
      tags:
      - user
      operationId: followUser
      description: 'Follow a user'
      parameters:
      - name: userID
        in: path
        required: true
        description: user id of the person in the session 
        schema:
          description: 'It sends which userID is requestion to follow a user'
          type: integer
          format: int64
      - name: followListOwnerID
        in: query
        required: true
        description: "Person I want to follow"
        schema:
          description: 'It sends userID that I want to follow'
          type: integer
          format: int64

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowRequest'
            example:
                userID: 123

      responses:
        # '201':
        #   description: User followed successfully
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '204':
          $ref: '#/components/responses/NoContentResponse'

    delete:
      security:
        - bearerAuth: []
      summary: Unfollow a user
      tags:
      - user
      operationId: unfollowUser
      description: 'un follow a user'
      parameters:
        - name: userID
          in: path
          required: true
          description: "Follower's user ID"
          schema:
            description: 'It sends userID that is follwoing'
            type: integer
            format: int64
        - name: followListOwnerID
          in: query
          required: true
          description: "User ID of the person to unfollow"
          schema:
            description: 'It sends userID that is being followed'
            type: integer
            format: int64
      responses:
        #no response for 200 
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '204':
          $ref: '#/components/responses/NoContentResponse'
  /users/{userID}/ban:
    post:
      security:
        - bearerAuth: []
      summary: Ban a user
      tags:
      - user
      operationId:  banUser
      description: 'Ban a user'
      parameters:
        - name: userID
          in: path
          required: true
          description: "User ID of the person in the seeion right now"
          schema:
            description: 'User who is in the session'
            type: integer
            format: int64
        - name: BannedUserID
          in: query
          required: true
          description: "User ID to be banned"
          schema:
            description: 'User ID to ban'
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BanRequest'
            example:
                userID: 123

      responses:
        # '200':
        #   description: User banned successfully
        # '201':
        #   description: User banned successfully
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '204':
          $ref: '#/components/responses/NoContentResponse'

    delete:
      security:
        - bearerAuth: []
      summary: Unban a user
      tags:
      - user
      operationId: unbanUser
      description: 'un ban a banned user'
      parameters:
        - name: userID
          in: path
          required: true
          description: "User ID to be unbanned"
          schema:
            description: 'User ID who has been banned'
            type: integer
            format: int64
            
        - name: BannedUserID #is the UserID of the person I want to ban
          in: query
          required: true
          description: "User ID of the person to ban"
          schema:
            description: 'User ID of the person in the session'
            type: integer
            format: int64

          
      responses:
        # '200':
        #   description: User unbanned successfully
        
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '204':
          $ref: '#/components/responses/NoContentResponse'
...
  /users/{userID}/image/{imageID}/comment:
    post:
      security:
        - bearerAuth: []
      summary: Add a comment to an image
      tags:
      - image
      operationId: commentPhoto
      description: 'Add comment to a user photo '
      parameters:
        - name: userID
          in: path
          required: true
          description: "User ID adding a comment"
          schema:
            description: "User ID who is adding comment"
            type: integer
            format: int64
        - name: imageID
          in: path
          required: true
          description: "Image ID to comment on"
          schema:
            description: "Image ID aka the image where the comment is being added"
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
            example:
              text: 'Nice picture'
      responses:
        # '201':
        #   description: Comment added successfully
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'

    delete:
      security:
        - bearerAuth: []
      summary: Remove a comment from an image
      tags:
      - image
      operationId: uncommentPhoto
      description: 'Remove a comment from a pic'
      parameters:
        - name: userID
          in: path
          required: true
          description: User ID removing a comment
          schema:
            description: "User ID of the person removing a comment"
            type: integer
        - name: imageID
          in: path
          required: true
          description: Image ID to remove the comment from
          schema:
            description: "Image ID to which the comment is at"
            type: integer
            format: int64
      responses:
      
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '204':
          $ref: '#/components/responses/NoContentResponse'
  /like/users/{userID}/image/{imageID}:
    post:
      security:
        - bearerAuth: []
      summary: Add a like to an image
      tags:
      - image
      operationId: likePhoto
      description: ' Add a like to a pic'
      parameters:
        - name: userID
          in: path
          required: true
          description: User ID adding a like
          schema:
            description: "User ID who is adding a like"
            type: integer
            format: int64
        - name: imageID
          in: path
          required: true
          description: Image ID to like
          schema:
            description: "Image ID to which the like is being added"
            type: integer
            format: int64
      responses:
        # '201':
        #   description: Like added successfully
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '204':
          $ref: '#/components/responses/NoContentResponse'
    delete:
      security:
        - bearerAuth: []
      summary: Remove a like from an image
      tags:
      - image
      operationId: unlikePhoto
      description: 'remove a like from a pic'
      parameters:
        - name: userID
          in: path
          required: true
          description: User ID removing a like
          schema:
            description: "UserID of the person remoing its like"
            type: integer
            format: int64
        - name: imageID
          in: path
          required: true
          description: Image ID to remove the like from
          schema:
            description: "Image ID aka the image to which the like is linked to"
            type: integer
            format: int64
      responses:
        # '200':
        #   description: Like removed successfully
        '404':
          $ref: '#/components/responses/NotFoundResponseUser'
        '204':
          $ref: '#/components/responses/NoContentResponse'
...